Implemented erasing of elements for company list call
The company list call produces a list of companies. In the event
of say 2 calls to this function. say 1st call gives a longer list (100 items)
than the 2nd call (20 items). The items in longer list doesnt get deleted. This
gives confusion as to which items are new and which are old. Not this patch deletes
items at the end until it finds blank space.

Tested and works!